works with controllers some permissions added

+++ app/Http/Controllers/PictureController.php
@@ -2,16 +2,20 @@
 
 namespace App\Http\Controllers;
 
+use App\Notifications\PictureChanged;
 use App\Picture;
+use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use JWTAuth;
 
 class PictureController extends Controller
 {
 
-    /*public function __construct()
+    public function __construct()
     {
         $this->middleware('jwt.auth',['except' => ['index','show']]);
-    }*/
+    }
     /**
      * Display a listing of the resource.
      *
@@ -48,16 +52,34 @@ class PictureController extends Controller
      */
     public function store(Request $request)
     {
+        $user = JWTAuth::parseToken()->authenticate();
+
+        $this->authorize('create', 'App\Picture');
+
         $picture = new Picture;
 
         $data['description'] = $request->description;
-        $data['user_id'] = $request->user_id;
+        $data['user_id'] = $user->id;
         $data['location'] = $request->location;
         $data['cost'] = $request->cost;
         $data['name'] = $request->name;
         $data['likes'] = $request->likes;
 
-        $picture = $picture->create($data);
+        if(!$picture = $picture->create($data)){
+
+            return response()->json(['message'=>'server error'],500);
+        }
+
+        if($followers = $user->followers()->get()){
+
+            foreach ($followers as $follower){
+
+                $follower->notify(new PictureChanged($picture));
+            }
+        }
+
+        $user->uploads++;
+        $user->save();
 
         return response()->json(['picture' => $picture], 200);
     }
@@ -99,20 +121,24 @@ class PictureController extends Controller
      */
     public function update(Request $request, Picture $picture)
     {
+        $this->authorize('update', $picture);
+
         $content = Picture::find($picture->id);
+
+        $user = JWTAuth::parseToken()->authenticate();
 
         if(!$content){
             return response()->json(['message'=>'Document not found'],404);
         }
 
         $data['description'] = $request->description;
-        $data['user_id'] = $request->user_id;
+        $data['user_id'] = $user->id;
         $data['location'] = $request->location;
         $data['cost'] = $request->cost;
         $data['name'] = $request->name;
         $data['likes'] = $request->likes;
 
-        $content->update($data);
+        $content->save($data);
 
         return response()->json(['picture' => $content],200);
 
@@ -126,8 +152,41 @@ class PictureController extends Controller
      */
     public function destroy(Picture $picture)
     {
+        $this->authorize('delete', $picture);
+
         $picture->delete();
+
+        $user = JWTAuth::parseToken()->authenticate();
+
+        $user->uploads = $user->uploads-1;
 
         return response()->json(['message'=>'picture deleted'],200);
     }
+
+    public function addToWatchlist(Picture $picture){
+
+        //get auth user $user
+        $user = JWTAuth::parseToken()->authenticate();
+
+        $this->authorize('addToWatchlist', 'App\Picture');
+
+        $user->watchlist()->attach($picture->id);
+
+        return response()->json(['message'=>'picture was added to watch list'],200);
+
+    }
+
+    public function removeFromWatchlist(Picture $picture){
+
+        $user = JWTAuth::parseToken()->authenticate();
+
+        //$this->authorize('watchlistCancel', $picture);
+
+        $user->watchlist()->detach($picture->id);
+
+        return response()->json(['message'=>'picture was removed from watch list'],200);
+
+    }
+
+
 }
